+++ front/src/components/slider/Slider.jsx
@@ -1,98 +1,145 @@
 import React from "react";
-import PropTypes from "prop-types";
 import { Link } from "react-router-dom";
-import { Navigation, Scrollbar, A11y } from "swiper";
+import { Autoplay } from "swiper";
 import { Swiper, SwiperSlide } from "swiper/react";
+
 import "swiper/scss";
 import "swiper/scss/navigation";
 import "swiper/scss/pagination";
+import img1 from "../../assets/images/box-item/img_item1.png";
+import img2 from "../../assets/images/box-item/img_item2.png";
+import img3 from "../../assets/images/box-item/img_item3.png";
 import shape1 from "../../assets/images/backgroup-secsion/bg-gradient1.png";
 import shape2 from "../../assets/images/backgroup-secsion/bg-gradient2.png";
 import shape3 from "../../assets/images/backgroup-secsion/bg-gradient3.png";
-import imgbg from "../../assets/images/backgroup-secsion/img_bg_page_title.jpg";
 
-const Slider = (props) => {
-  const data = props.data;
+const SliderStyle4 = () => {
   return (
-    <div className="mainslider">
-      <Swiper
-        modules={[Navigation, Scrollbar, A11y]}
-        spaceBetween={0}
-        slidesPerView={1}
-        navigation
-        scrollbar={{ draggable: true }}
-      >
-        {data.map((item, index) => (
-          <SwiperSlide key={index} className={item.class}>
-            <SliderItem item={item} />
-          </SwiperSlide>
-        ))}
-      </Swiper>
+    <div>
+      <section className="flat-title-page style3 mainslider">
+        <img className="bgr-gradient gradient1" src={shape1} alt="Ticketing" />
+        <img className="bgr-gradient gradient2" src={shape2} alt="Ticketing" />
+        <img className="bgr-gradient gradient3" src={shape3} alt="Ticketing" />
+        <div className="overlay"></div>
+        <div className="themesflat-container ">
+          <div className="wrap-heading flat-slider flex">
+            <div className="content">
+              <h2 className="heading mt-15">Discover, find,</h2>
+              <h1 className="heading mb-style">
+                <span className="tf-text s1">Sell extraordinary</span>
+              </h1>
+              <h1 className="heading">Monster Tickets</h1>
+              <p className="sub-heading mt-19 mb-40">
+                Marketplace for monster character cllections non fungible token
+                Tickets
+              </p>
+              <div className="flat-bt-slider flex style2">
+                <Link
+                  to="/explore"
+                  className="sc-button header-slider style style-1 rocket fl-button pri-1"
+                >
+                  <span>Explore</span>
+                </Link>
+                <Link
+                  to="/create-item"
+                  className="sc-button header-slider style style-1 note fl-button pri-1"
+                >
+                  <span>Create</span>
+                </Link>
+              </div>
+            </div>
+            <Swiper
+              modules={[Autoplay]}
+              direction={"vertical"}
+              spaceBetween={25}
+              slidesPerView={5}
+              loop
+              autoplay={{
+                delay: 1,
+                disableOnInteraction: false,
+                pauseOnMouseEnter: true,
+              }}
+              speed={2000}
+            >
+              <SwiperSlide>
+                <img src={img1} alt="Ticketing" />
+              </SwiperSlide>
+              <SwiperSlide>
+                <img src={img2} alt="Ticketing" />
+              </SwiperSlide>
+              <SwiperSlide>
+                <img src={img3} alt="Ticketing" />
+              </SwiperSlide>
+              <SwiperSlide>
+                <img src={img2} alt="Ticketing" />
+              </SwiperSlide>
+              <SwiperSlide>
+                <img src={img3} alt="Ticketing" />
+              </SwiperSlide>
+            </Swiper>
+            <Swiper
+              modules={[Autoplay]}
+              direction={"vertical"}
+              spaceBetween={25}
+              slidesPerView={5}
+              loop
+              autoplay={{
+                delay: 1,
+                disableOnInteraction: false,
+                pauseOnMouseEnter: true,
+              }}
+              speed={1800}
+            >
+              <SwiperSlide>
+                <img src={img3} alt="Ticketing" />
+              </SwiperSlide>
+              <SwiperSlide>
+                <img src={img2} alt="Ticketing" />
+              </SwiperSlide>
+              <SwiperSlide>
+                <img src={img3} alt="Ticketing" />
+              </SwiperSlide>
+              <SwiperSlide>
+                <img src={img1} alt="Ticketing" />
+              </SwiperSlide>
+              <SwiperSlide>
+                <img src={img2} alt="Ticketing" />
+              </SwiperSlide>
+            </Swiper>
+            <Swiper
+              modules={[Autoplay]}
+              direction={"vertical"}
+              spaceBetween={25}
+              slidesPerView={5}
+              loop
+              autoplay={{
+                delay: 1,
+                disableOnInteraction: false,
+                pauseOnMouseEnter: true,
+              }}
+              speed={2200}
+            >
+              <SwiperSlide>
+                <img src={img2} alt="Ticketing" />
+              </SwiperSlide>
+              <SwiperSlide>
+                <img src={img1} alt="Ticketing" />
+              </SwiperSlide>
+              <SwiperSlide>
+                <img src={img3} alt="Ticketing" />
+              </SwiperSlide>
+              <SwiperSlide>
+                <img src={img2} alt="Ticketing" />
+              </SwiperSlide>
+              <SwiperSlide>
+                <img src={img1} alt="Ticketing" />
+              </SwiperSlide>
+            </Swiper>
+          </div>
+        </div>
+      </section>
     </div>
   );
 };
 
-Slider.propTypes = {
-  data: PropTypes.array.isRequired,
-  control: PropTypes.bool,
-  auto: PropTypes.bool,
-  timeOut: PropTypes.number,
-};
-const SliderItem = (props) => (
-  <div className="flat-title-page" style={{ backgroundImage: `url(${imgbg})` }}>
-    <img className="bgr-gradient gradient1" src={shape1} alt="Ticketing" />
-    <img className="bgr-gradient gradient2" src={shape2} alt="Ticketing" />
-    <img className="bgr-gradient gradient3" src={shape3} alt="Ticketing" />
-    <div className="shape item-w-16"></div>
-    <div className="shape item-w-22"></div>
-    <div className="shape item-w-32"></div>
-    <div className="shape item-w-48"></div>
-    <div className="shape style2 item-w-51"></div>
-    <div className="shape style2 item-w-51 position2"></div>
-    <div className="shape item-w-68"></div>
-    <div className="overlay"></div>
-    <div className="swiper-container mainslider home">
-      <div className="swiper-wrapper">
-        <div className="swiper-slide">
-          <div className="slider-item">
-            <div className="themesflat-container ">
-              <div className="wrap-heading flat-slider flex">
-                <div className="content">
-                  <h2 className="heading">{props.item.title_1}</h2>
-                  <h1 className="heading mb-style">
-                    <span className="tf-text s1">{props.item.title_2}</span>
-                  </h1>
-                  <h1 className="heading">{props.item.title_3}</h1>
-                  <p className="sub-heading">{props.item.description}</p>
-                  <div className="flat-bt-slider flex style2">
-                    <Link
-                      to="/explore"
-                      className="sc-button header-slider style style-1 rocket fl-button pri-1"
-                    >
-                      <span>Explore</span>
-                    </Link>
-                    <Link
-                      to="/create-item"
-                      className="sc-button header-slider style style-1 note fl-button pri-1"
-                    >
-                      <span>Create</span>
-                    </Link>
-                  </div>
-                </div>
-                <div className="image">
-                  <img
-                    className="img-bg"
-                    src={props.item.imgbg}
-                    alt="ticketing"
-                  />
-                  <img src={props.item.img} alt="ticketing" />
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-);
-export default Slider;
+export default SliderStyle4;
