+++ front/src/pages/Home.jsx
@@ -1,10 +1,10 @@
 import React from "react";
-import HeaderStyle2 from "../components/header/HeaderStyle2";
+import HeaderHome from "../components/header/HeaderHome";
 import Footer from "../components/footer/Footer";
 import Category from "../components/layouts/home/Category";
 import LiveAuction from "../components/layouts/home/LiveAuction";
 import TopSeller from "../components/layouts/home/TopSeller";
-import SliderStyle4 from "../components/slider/SliderStyle4";
+import Slider from "../components/slider/Slider";
 import TodayPicks from "../components/layouts/home/TodayPicks";
 import Create from "../components/layouts/home/Create";
 import PopularCollection from "../components/layouts/home/PopularCollection";
@@ -12,8 +12,8 @@ import PopularCollection from "../components/layouts/home/PopularCollection";
 const Home = () => {
   return (
     <div className="home">
-      <HeaderStyle2 />
-      <SliderStyle4 />
+      <HeaderHome />
+      <Slider />
       <LiveAuction />
       <Category />
       <TopSeller />
